@@ -10,10 +10,14 @@ import java.util.HashSet;
 import java.util.Set;
 
 import com.momoko.es.api.dto.AutorDTO;
+import com.momoko.es.api.dto.EditorialDTO;
+import com.momoko.es.api.dto.GeneroDTO;
 import com.momoko.es.api.dto.LibroDTO;
 import com.momoko.es.api.dto.UsuarioDTO;
 import com.momoko.es.api.enums.UserStatusEnum;
 import com.momoko.es.model.entity.AutorEntity;
+import com.momoko.es.model.entity.EditorialEntity;
+import com.momoko.es.model.entity.GeneroEntity;
 import com.momoko.es.model.entity.LibroEntity;
 import com.momoko.es.model.entity.UsuarioEntity;
 
@@ -77,9 +81,9 @@ public final class DTOToEntityAdapter {
         libroEntity.setAnoEdicion(libroDTO.getAnoEdicion());
         libroEntity.setAutores(adaptarAutores(libroDTO.getAutores()));
         libroEntity.setCitaLibro(libroDTO.getCitaLibro());
-        libroEntity.setEditorialId(libroDTO.getEditorialId());
+        libroEntity.setEditorial(adaptarEditorial(libroDTO.getEditorial()));
         libroEntity.setEnlaceAmazon(libroDTO.getEnlaceAmazon());
-        libroEntity.setGeneroId(libroDTO.getGeneroId());
+        libroEntity.setGeneros(adaptarGeneros(libroDTO.getGeneros()));
         libroEntity.setLibroId(libroDTO.getLibroId());
         libroEntity.setResumen(libroDTO.getResumen());
         libroEntity.setSagaId(libroDTO.getSagaId());
@@ -88,7 +92,28 @@ public final class DTOToEntityAdapter {
         return libroEntity;
     }
 
-    private static Set<AutorEntity> adaptarAutores(final Set<AutorDTO> autores) {
+    /**
+     * Adaptar editorial.
+     *
+     * @param editorial
+     *            the editorial
+     * @return the editorial entity
+     */
+    private static EditorialEntity adaptarEditorial(final EditorialDTO editorial) {
+        final EditorialEntity editorialEntity = new EditorialEntity();
+        editorialEntity.setEditorialId(editorial.getEditorialId());
+        editorialEntity.setNombreEditorial(editorial.getNombreEditorial());
+        return editorialEntity;
+    }
+
+    /**
+     * Adaptar autores.
+     *
+     * @param autores
+     *            the autores
+     * @return the establece
+     */
+    public static Set<AutorEntity> adaptarAutores(final Set<AutorDTO> autores) {
         final Set<AutorEntity> autoresEntities = new HashSet<AutorEntity>();
         for (final AutorDTO autorDTO : autores) {
             autoresEntities.add(adaptarAutor(autorDTO));
@@ -96,10 +121,46 @@ public final class DTOToEntityAdapter {
         return autoresEntities;
     }
 
-    private static AutorEntity adaptarAutor(final AutorDTO autorDTO) {
+    /**
+     * Adaptar autor.
+     *
+     * @param autorDTO
+     *            the autor dto
+     * @return the autor entity
+     */
+    public static AutorEntity adaptarAutor(final AutorDTO autorDTO) {
         final AutorEntity entity = new AutorEntity();
         entity.setAutorId(autorDTO.getAutorId());
         entity.setNombre(autorDTO.getNombre());
+        return entity;
+    }
+
+    /**
+     * Adaptar autores.
+     *
+     * @param autores
+     *            the autores
+     * @return the establece
+     */
+    public static Set<GeneroEntity> adaptarGeneros(final Set<GeneroDTO> generos) {
+        final Set<GeneroEntity> generosEntities = new HashSet<GeneroEntity>();
+        for (final GeneroDTO generoDTO : generos) {
+            generosEntities.add(adaptarGenero(generoDTO));
+        }
+        return generosEntities;
+    }
+
+    /**
+     * Adaptar genero.
+     *
+     * @param generoDTO
+     *            the genero dto
+     * @return the genero entity
+     */
+    public static GeneroEntity adaptarGenero(final GeneroDTO generoDTO) {
+        final GeneroEntity entity = new GeneroEntity();
+        entity.setGeneroId(generoDTO.getGeneroId());
+        entity.setNombre(generoDTO.getNombre());
         return entity;
     }
 }
